requirements.txt
streamlit==1.32.0
supabase==1.0.3
python-dotenv==1.0.1
asyncpg
Pillow==10.2.0
pandas==2.2.1
python-dateutil==2.8.2
requests==2.31.0
python-multipart==0.0.9
uuid==1.30
postgrest==0.10.7
gotrue==1.3.1
httpx==0.23.3
storage3==0.5.3
supafunc==0.2.2
aiohttp==3.9.3
asyncio==3.4.3
ujson
country-list
streamlit-extras 